Create DayFilter class and separate its usage from ForecastFilter
