+++ src/main/java/athena/TimeAllocator.java
@@ -5,6 +5,7 @@ import athena.exceptions.DateHasPassedException;
 import athena.exceptions.TaskNotFoundException;
 import athena.task.Task;
 import athena.task.Time;
+import athena.task.taskfilter.DayFilter;
 import athena.task.taskfilter.FlexibleTimeFilter;
 import athena.task.taskfilter.ForecastFilter;
 
@@ -124,8 +125,8 @@ public class TimeAllocator {
 
 
     private TaskList getFixedDayTasks(LocalDate date) {
-        ForecastFilter forecast = new ForecastFilter(date);
-        TaskList fixedDayTask = this.fixedTaskList.getFilteredList(forecast);
+        DayFilter filter = new DayFilter(date);
+        TaskList fixedDayTask = this.fixedTaskList.getFilteredList(filter);
         return fixedDayTask;
     }
