+++ src/main/java/athena/task/taskfilter/ForecastFilter.java
@@ -1,6 +1,7 @@
 package athena.task.taskfilter;
 
 import athena.Forecast;
+import athena.common.utils.DateUtils;
 import athena.task.Task;
 
 import java.time.LocalDate;
@@ -17,11 +18,6 @@ public class ForecastFilter extends TaskFilter {
         this.forecast = forecast;
     }
 
-    public ForecastFilter(LocalDate date) {
-        this.forecast = Forecast.DAY;
-        this.filterDate = date;
-    }
-
     public void setDate(LocalDate filterDate) {
         this.filterDate = filterDate;
     }
@@ -35,29 +31,10 @@ public class ForecastFilter extends TaskFilter {
     @Override
     public boolean isTaskIncluded(Task task) {
         for (LocalDate date : task.getDates()) {
-            if (isDateIncluded(date)) {
+            if (DateUtils.isDateIncluded(date, forecast)) {
                 return true;
             }
         }
         return false;
-    }
-
-    private static int getWeekNumber(LocalDate taskDate) {
-        TemporalField woy = WeekFields.of(Locale.getDefault()).weekOfWeekBasedYear();
-        return taskDate.get(woy);
-    }
-
-    private boolean isDateIncluded(LocalDate taskDate) {
-        boolean isDateIncluded;
-        if (forecast == Forecast.ALL) {
-            isDateIncluded = true;
-        } else if (forecast == Forecast.WEEK) {
-            int currentWeekNumber = getWeekNumber(filterDate);
-            int taskWeekNumber = getWeekNumber(taskDate);
-            isDateIncluded = (currentWeekNumber == taskWeekNumber && taskDate.getYear() == LocalDate.now().getYear());
-        } else {
-            isDateIncluded = taskDate.equals(filterDate);
-        }
-        return isDateIncluded;
     }
 }
